+++ CHANGES.md
@@ -4,6 +4,7 @@
 Changes in 11.4.4
 -----------------
 - Fix SQL transcript_grades view for History Marking Periods of existing school years in rosariosis.sql, rosariosis_mysql.sql & Update.fnc.php
+- Fix regression since 11.4 check MP exists before trying to get GPA: include History Marking Periods in Grades.fnc.php
 
 Changes in 11.4.3
 -----------------